split up auth tests
