+++ apps/playwright/constants/auth.ts
@@ -7,4 +7,66 @@ export default {
     COMMUNITY_ADMIN: ".auth/community-admin.json",
     PROCESS_OPERATOR: ".auth/process-operator.json",
   },
+  RESTRICTED_PATHS: {
+    ADMIN: [
+      "/en/admin",
+      "/en/admin/talent-requests",
+      "/en/admin/users",
+      "/en/admin/settings/classifications",
+      "/en/admin/pools",
+      "/en/admin/settings/departments",
+      "/en/admin/settings/skills",
+      "/en/admin/settings/skill-families",
+    ],
+    COMMUNITY_ADMIN: [
+      "/en/admin",
+      "/en/admin/settings/announcements",
+      "/en/admin/settings/classifications",
+      "/en/admin/settings/departments",
+      "/en/admin/settings/skills",
+      "/en/admin/settings/skill-families",
+    ],
+    COMMUNITY_RECRUITER: [
+      "/en/admin",
+      "/en/admin/settings/announcements",
+      "/en/admin/settings/classifications",
+      "/en/admin/settings/departments",
+      "/en/admin/settings/skills",
+      "/en/admin/settings/skill-families",
+    ],
+    PROCESS_OPERATOR: [
+      "/en/admin",
+      "/en/admin/settings/announcements",
+      "/en/admin/settings/classifications",
+      "/en/admin/settings/departments",
+      "/en/admin/settings/skills",
+      "/en/admin/settings/skill-families",
+      "/en/admin/talent-requests",
+      "/en/admin/communities",
+    ],
+  },
+  ALLOWED_PATHS: {
+    COMMUNITY_ADMIN: [
+      "/en/community",
+      "/en/admin/pools",
+      "/en/admin/pool-candidates",
+      "/en/admin/talent-requests",
+      "/en/admin/communities",
+      "/en/admin/users",
+    ],
+    COMMUNITY_RECRUITER: [
+      "/en/community",
+      "/en/admin/pools",
+      "/en/admin/pool-candidates",
+      "/en/admin/talent-requests",
+      "/en/admin/communities",
+      "/en/admin/users",
+    ],
+    PROCESS_OPERATOR: [
+      "/en/community",
+      "/en/admin/users",
+      "/en/admin/pools",
+      "/en/admin/pool-candidates",
+    ],
+  },
 };
